+++ weblate/formats/base.py
@@ -22,7 +22,9 @@ from translate.storage.base import TranslationUnit as TranslateToolkitUnit
 from weblate_language_data.countries import DEFAULT_LANGS
 
 from weblate.checks.flags import Flags
-from weblate.trans.file_format_params import get_params_for_file_format
+from weblate.trans.file_format_params import (
+    get_params_for_file_format,
+)
 from weblate.trans.util import get_string, join_plural, split_plural
 from weblate.utils.errors import add_breadcrumb
 from weblate.utils.files import get_repo_temp_dir
@@ -37,7 +39,9 @@ if TYPE_CHECKING:
     from lxml import etree
 
     from weblate.lang.models import Language, Plural
-    from weblate.trans.file_format_params import FileFormatParams
+    from weblate.trans.file_format_params import (
+        FileFormatParams,
+    )
     from weblate.trans.models import Component, Translation, Unit
     from weblate.utils.state import StringState
 
@@ -567,7 +571,7 @@ class TranslationFormat[S: InnerStore, U: InnerUnit, T: TranslationUnit]:
         """Add new unit to underlying store."""
         raise NotImplementedError
 
-    def update_header(self, **kwargs) -> None:
+    def update_header(self, file_format_params: FileFormatParams, **kwargs) -> None:
         """Update store header if available."""
         return
 
@@ -763,7 +767,7 @@ class TranslationFormat[S: InnerStore, U: InnerUnit, T: TranslationUnit]:
     def add_language(
         cls,
         filename: str | Path,
-        language: str,
+        language: Language,
         base: str,
         callback: Callable | None = None,
         file_format_params: FileFormatParams | None = None,
@@ -792,7 +796,7 @@ class TranslationFormat[S: InnerStore, U: InnerUnit, T: TranslationUnit]:
     def create_new_file(
         cls,
         filename: str,
-        language: str,
+        language: Language,
         base: str,
         callback: Callable | None = None,
         file_format_params: FileFormatParams | None = None,
@@ -1032,6 +1036,7 @@ class BaseExporter:
     set_id = False
     file_format = ""
     storage_class: ClassVar[type[TranslateToolkitStore]]
+    file_format_params: FileFormatParams
 
     def __init__(
         self,
@@ -1049,12 +1054,16 @@ class BaseExporter:
             self.source_language = translation.component.source_language
             self.language = translation.language
             self.url = get_site_url(translation.get_absolute_url())
+            self.file_format_params = cast(
+                "FileFormatParams", self.translation.component.file_format_params
+            )
         else:
             self.project = project
             self.language = language
             self.source_language = source_language
             self.plural = language.plural
             self.url = url
+            self.file_format_params = cast("FileFormatParams", {})
         self.fieldnames = fieldnames
 
     @staticmethod
@@ -1088,13 +1097,12 @@ class BaseExporter:
         if self.translation is None or not self.file_format:
             return
 
-        params = self.translation.component.file_format_params
         for param_class in get_params_for_file_format(self.file_format):
             if param_class.is_encoding():
-                encoding = param_class.get_value(params)
+                encoding = param_class.get_value(self.file_format_params)
                 if encoding != "auto":
                     storage.encoding = encoding
-            param_class().setup_store(storage, **params)
+            param_class().setup_store(storage, **self.file_format_params)
 
     def add(self, unit: TranslateToolkitUnit, word: str) -> None:
         unit.target = word