+++ weblate/trans/backups.py
@@ -138,6 +138,8 @@ class ProjectBackup:
         self.components_cache: dict[str, Component] = {}
         self.categories_cache: dict[str, Category] = {}
         self.roles_cache: dict[str, Role] = {}
+        # Project.set_language_team field was migrated to file format parameters after 5.17
+        self.set_language_team_project: bool = False
 
     @staticmethod
     def full_slug_without_project(obj: Component | Category) -> str:
@@ -921,6 +923,15 @@ class ProjectBackup:
             using=None,
             update_fields=None,
         )
+
+        if component.file_format in {"po", "po-mono"} and (
+            "file_format_params" not in kwargs
+            or kwargs["file_format_params"].get("po_set_language_team") is None
+        ):
+            # fallback to project setting if not set in backup
+            component.file_format_params["po_set_language_team"] = (
+                self.set_language_team_project
+            )
         # Use bulk create to avoid triggering save() and any post_save signals
         component = Component.objects.bulk_create([component])[0]
 
@@ -1148,6 +1159,8 @@ class ProjectBackup:
             kwargs = self.data["project"].copy()
             kwargs["name"] = project_name
             kwargs["slug"] = project_slug
+            # the attribute `set_language_team` is present in legacy backups prior to 5.17.1
+            self.set_language_team_project = kwargs.pop("set_language_team", False)
             self.project = project = Project.objects.create(**kwargs)
 
             # Handle billing and ACL (creating user needs access)