+++ weblate/trans/file_format_params.py
@@ -36,6 +36,10 @@ class FileFormatParams(TypedDict, total=False):
     po_keep_previous: bool
     po_no_location: bool
     po_fuzzy_matching: bool
+    po_set_language_team: bool
+    po_set_last_translator: bool
+    po_set_x_generator: bool
+    po_report_msgid_bugs_to: bool
     yaml_indent: int
     yaml_line_wrap: int
     yaml_line_break: str
@@ -61,6 +65,10 @@ class BaseFileFormatParam:
         "po_keep_previous",
         "po_no_location",
         "po_fuzzy_matching",
+        "po_set_language_team",
+        "po_set_last_translator",
+        "po_set_x_generator",
+        "po_report_msgid_bugs_to",
         "yaml_indent",
         "yaml_line_wrap",
         "yaml_line_break",
@@ -75,7 +83,7 @@ class BaseFileFormatParam:
         "gwt_encoding",
         "merge_duplicates",
     ]
-    file_formats: Sequence[str] = []
+    file_formats: Sequence[str] = ()
     field_class: type[forms.Field] = forms.CharField
     label: StrOrPromise = ""
     default: str | int | bool
@@ -120,7 +128,9 @@ class BaseFileFormatParam:
         """Configure store with this file format parameters."""
 
     @classmethod
-    def get_value(cls, file_format_params: FileFormatParams):
+    def get_value(cls, file_format_params: FileFormatParams | None):
+        if file_format_params is None:
+            return cls.default
         value = file_format_params.get(cls.name, cls.default)
         if value is None:
             return cls.default
@@ -315,7 +325,7 @@ class GettextPoLineWrap(BaseFileFormatParam):
 
 
 class BaseGettextFormatParam(BaseFileFormatParam):
-    file_formats = ("po",)
+    file_formats: Sequence[str] = ("po",)
 
 
 @register_file_format_param
@@ -340,6 +350,48 @@ class GettextFuzzyMatching(BaseGettextFormatParam):
     label = gettext_lazy("Use fuzzy matching")
     field_class = forms.BooleanField
     default = True
+
+
+@register_file_format_param
+class GettextSetLanguageTeamHeader(BaseGettextFormatParam):
+    file_formats = ("po", "po-mono")
+    name = "po_set_language_team"
+    label = gettext_lazy("Update language team header")
+    field_class = forms.BooleanField
+    default = True
+    help_text = gettext_lazy('Lets Weblate update the "Language-Team" file header.')
+
+
+@register_file_format_param
+class GettextLastTranslator(BaseGettextFormatParam):
+    file_formats = ("po", "po-mono")
+    name = "po_set_last_translator"
+    label = gettext_lazy("Update last translator header")
+    field_class = forms.BooleanField
+    default = True
+    help_text = gettext_lazy('Lets Weblate update the "Last-Translator" file header.')
+
+
+@register_file_format_param
+class GettextXGenerator(BaseGettextFormatParam):
+    file_formats = ("po", "po-mono")
+    name = "po_set_x_generator"
+    label = gettext_lazy("Update X-Generator header")
+    field_class = forms.BooleanField
+    default = True
+    help_text = gettext_lazy('Lets Weblate update the "X-Generator" file header.')
+
+
+@register_file_format_param
+class GettextReportMsgidBugsTo(BaseGettextFormatParam):
+    file_formats = ("po", "po-mono")
+    name = "po_report_msgid_bugs_to"
+    label = gettext_lazy("Report msgid bugs to")
+    field_class = forms.BooleanField
+    default = True
+    help_text = gettext_lazy(
+        'Lets Weblate update the "Report-Msgid-Bugs-To" file header if Source string bug reporting address is set.'
+    )
 
 
 class BaseYAMLFormatParam(BaseFileFormatParam):