+++ weblate/trans/forms.py
@@ -2459,7 +2459,6 @@ class ProjectSettingsForm(SettingsBaseForm, ProjectDocsMixin, ProjectAntispamMix
             "name",
             "web",
             "instructions",
-            "set_language_team",
             "use_shared_tm",
             "contribute_shared_tm",
             "autoclean_tm",
@@ -2607,7 +2606,6 @@ class ProjectSettingsForm(SettingsBaseForm, ProjectDocsMixin, ProjectAntispamMix
                 ),
                 Tab(
                     gettext("Workflow"),
-                    "set_language_team",
                     "use_shared_tm",
                     "contribute_shared_tm",
                     "autoclean_tm",