+++ weblate/trans/models/project.py
@@ -208,13 +208,6 @@ class Project(models.Model, PathMixin, CacheKeyMixin, LockMixin):
         help_text=gettext_lazy("You can use Markdown and mention users by @username."),
     )
 
-    set_language_team = models.BooleanField(
-        verbose_name=gettext_lazy('Set "Language-Team" header'),
-        default=True,
-        help_text=gettext_lazy(
-            'Lets Weblate update the "Language-Team" file header of your project.'
-        ),
-    )
     use_shared_tm = models.BooleanField(
         verbose_name=gettext_lazy("Use shared translation memory"),
         default=settings.DEFAULT_SHARED_TM,